Move data munging to lambda
