+++ .gitignore
@@ -5,3 +5,4 @@ static/css
 static/js
 dist
 .idea
+*.gz
