+++ charting_covid_data/chalicelib/create_covid_json.py
@@ -51,6 +51,50 @@ def get_case_count(output_data: dict, date: str, fips: str,
     return None
 
 
+def set_increase_count(output_data: dict, date: str, fips: str, increase: int,
+        output_fips_first: bool) -> None:
+    """
+    Sets the given increase count into output_data for a Date-FIPS combo.
+    Does this based on the output_fips_first option.
+    """
+    if output_fips_first:
+        # output_data is keyed by FIPS, then date
+        # {"56043":
+        #    {"2020-03-27": {"increase": 12}, ...}
+        # }
+        entry = output_data[fips][date]
+        entry["increase"] = increase
+    else:
+        # output_data is keyed by date, and then FIPS.
+        # {"2020-03-27":
+        #    {"56043": {"increase": 12}, ...}
+        # }
+        entry = output_data[date][fips]
+        entry["increase"] = increase
+
+
+def set_per_capita(output_data: dict, date: str, fips: str, cases: int,
+        population: int, output_fips_first: bool) -> None:
+    """
+    Sets the per capita case count into output_data for a Date-FIPS combo.
+    Does this based on the output_fips_first option.
+    """
+    if output_fips_first:
+        # output_data is keyed by FIPS, then date
+        # {"56043":
+        #    {"2020-03-27": {"per_capita": 0.12}, ...}
+        # }
+        entry = output_data[fips][date]
+        entry["per_capita"] = cases / population
+    else:
+        # output_data is keyed by date, and then FIPS.
+        # {"2020-03-27":
+        #    {"56043": {"per_capita": 0.12}, ...}
+        # }
+        entry = output_data[date][fips]
+        entry["per_capita"] = cases / population
+
+
 def get_exp_growth_rate(final: Num, starting: Num, num_periods: Num) -> float:
     """
     The exp function is Final = (Starting)e**(Rate*Periods)
@@ -72,7 +116,7 @@ def get_doubling_time(exp_growth_rate: float) -> float:
     return math.log(2) / math.log(1 + exp_growth_rate)
 
 
-def get_averaged_doubling_time(preceding_case_counts: list) -> float:
+def get_averaged_doubling_time(preceding_case_counts: list) -> Optional[float]:
     """
     The idea here is to return the average estimated doubling time of each of
     the preceding counts and the current count.
@@ -89,7 +133,9 @@ def get_averaged_doubling_time(preceding_case_counts: list) -> float:
             latest_count, earlier_count, num_periods=i)
         if exp_growth_rate > 0:
             growth_rates.append(exp_growth_rate)
-    return mean([get_doubling_time(r) for r in growth_rates])
+    if len(growth_rates):
+        return mean([get_doubling_time(r) for r in growth_rates])
+    return None
 
 
 def get_daily_increases(preceding_case_counts: list) -> list:
@@ -221,7 +267,7 @@ def record_growth_metrics(output_data: dict,
     return output_data
 
 
-def record_case_counts(csv_data: list, output_data: dict,
+def record_case_counts(csv_data: list, output_data: dict, fips_data: dict,
         inverse_chronological_case_data: dict, output_fips_first: bool,
         is_state_file: bool, latest_date: datetime.date) -> (dict, dict):
     """
@@ -309,10 +355,25 @@ def record_case_counts(csv_data: list, output_data: dict,
             for i in range(1, previous_days_adjusted_down + 1):
                 fips_row[day_offset + i] = cases
 
+        if fips_data:
+            # Set the per-capita case count.
+            population = fips_data.get(row[FIPS], {}).get("population")
+            if population:
+                set_per_capita(output_data, row[DATE], row[FIPS], cases,
+                    population, output_fips_first)
+
+        # Set this days case increase.
+        preceding_counts = \
+            inverse_chronological_case_data[row[FIPS]][day_offset:day_offset + 2]
+        if len(preceding_counts) == 2 and None not in preceding_counts:
+            increase = preceding_counts[0] - preceding_counts[1]
+            set_increase_count(output_data, row[DATE], row[FIPS], increase,
+                output_fips_first)
+
     return output_data, inverse_chronological_case_data
 
 
-def generate_covid_data(covid_data: list, output_data: dict,
+def generate_covid_data(covid_data: list, output_data: dict, fips_data: dict,
         growth_metric_days: int, output_fips_first: bool,
         is_state_file: bool=False) -> dict:
     """
@@ -357,8 +418,9 @@ def generate_covid_data(covid_data: list, output_data: dict,
         defaultdict(lambda: [None] * total_days_of_data)
 
     output_data, inverse_chronological_case_data = record_case_counts(
-            covid_data, output_data, inverse_chronological_case_data,
-            output_fips_first, is_state_file, latest_date)
+            covid_data, output_data, fips_data,
+            inverse_chronological_case_data, output_fips_first, is_state_file,
+            latest_date)
 
     output_data = record_growth_metrics(output_data,
             inverse_chronological_case_data, output_fips_first,
@@ -368,13 +430,13 @@ def generate_covid_data(covid_data: list, output_data: dict,
 
 
 def generate_case_json(county_input: list, state_input: list, output_file: str,
-        growth_metric_days: int) -> None:
+        fips_data: dict, growth_metric_days: int) -> None:
     empty_data = defaultdict(dict)
     state_data = generate_covid_data(
-        state_input, empty_data, growth_metric_days, False,
+        state_input, empty_data, fips_data, growth_metric_days, False,
         is_state_file=True)
     state_and_county_data = generate_covid_data(
-        county_input, state_data, growth_metric_days, False)
+        county_input, state_data, fips_data, growth_metric_days, False)
 
     json_str = json.dumps(state_and_county_data)
     json_bytes = json_str.encode("utf-8")
